// src/constants/game/super-wheel.enum.ts
export enum SuperWheelElement {
  ORANGE = "ORANGE",
  PLUM = "PLUM",
  PEAR = "PEAR",
  STRAWBERRY = "STRAWBERRY",
  GRAPE = "GRAPE",
  WATERMELON = "WATERMELON",
  BELL = "BELL",
  SEVEN = "SEVEN",
  KING = "KING",
  CARD = "CARD",
}

export type SuperWheel = SuperWheelElement | [SuperWheelElement, SuperWheelElement];